test6 refactor

+++ exercise.main/Basket.cs
@@ -44,6 +44,6 @@ namespace exercise.main
 
         public bool IsFull { get { return _capacity == _products.Count(); } }
 
-        public int SumOfItems { get { return ((int)_products.Sum(product => product.Price)); } }
+        public double SumOfItems { get { return (_products.Sum(product => product.Price)); } }
     }
 }
